Chef's choice
The chefs can choose what they want to add for some dishes. This is mostly random right now, but I will either make it so each choice of thing to add is its own action or a mini plan is made to determine which choice is the best.
Added some new recipes and a meals list.